+++ src/validators/string/index.ts
@@ -2,10 +2,12 @@
 export { alpha } from './alpha';
 export { alphanumeric } from './alphanumeric';
 export { chars } from './chars';
+export { differentFrom } from './different-from';
 export { email } from './email';
 export { maxChars } from './max-chars';
 export { minChars } from './min-chars';
 export { numeric } from './numeric';
+export { password } from './password';
 export { phone } from './phone';
 export { regex } from './regex';
 export { required } from './required';